refactor: 모바일에서 넘치던 모달 컴포넌트 교체하여 해결 및 원인 컴포넌트 삭제
